fix: fix issue #15
Fixed by creating a getCanvasResolution() and returning the values of
canvasWidth and canvasHeight of canvasScaler.js as an object. For now
these two variables initiated by calling the initializeZoomButtons()
function are the current reliable way of getting the initial canvas
resolution required for this issue.

The returned values of getCanvasResolution() is then used as values
for DOM input boxes with id's changeResWidth and changeResHeight.
Currently this makes the function changeResInitialValues() from
canvasResolution.js obsolete.